fix(clippy): batch 1 - machine-applicable fixes + blas_openblas disambiguation
- apply machine-applicable clippy suggestions across 54 files (uninlined
  format args, redundant patterns, manual_* idioms, etc.); behavior-preserving
- fix pre-existing E0034 under --all-features: disambiguate OpenBlasProvider
  name() call in tests via fully-qualified BlasProviderF32::name(&p)

Validated: default-feature workspace builds + tests pass (pre-existing STM
producer/consumer flake aside). --all-features test linking needs system
openblas (not available locally; clippy type-checks fine without it).

+++ crates/bhc-diagnostics/src/lsp.rs
@@ -299,12 +299,12 @@ pub fn span_to_range(file: &SourceFile, span: bhc_span::Span) -> LspRange {
     LspRange {
         start: LspPosition {
             // LSP uses 0-indexed lines
-            line: start_loc.line.saturating_sub(1) as u32,
-            character: start_loc.col.saturating_sub(1) as u32,
+            line: start_loc.line.saturating_sub(1),
+            character: start_loc.col.saturating_sub(1),
         },
         end: LspPosition {
-            line: end_loc.line.saturating_sub(1) as u32,
-            character: end_loc.col.saturating_sub(1) as u32,
+            line: end_loc.line.saturating_sub(1),
+            character: end_loc.col.saturating_sub(1),
         },
     }
 }
